BP Calculator Fix and Enhancement  * Fixed calculation of ME Bonus for materials on a BP  * Switched runs variables to a long from an int.  Int was overflowing on larger BPs causing a crash.  * Added a 'Station ME Bonus' control on BP Calculator screen.  Allows user to factor in ME bonus that some null sec stations provide.

+++ EveHQ.Core/SkillFunctions.vb
@@ -52,11 +52,11 @@ Public Class SkillFunctions
     ' Shared variables for repeated usage
     Shared sTimeSpan, eTimeSpan, tTimeSpan As TimeSpan
 
-    Public Shared Function Roman(ByVal dec As Integer) As String
+	Public Shared Function Roman(ByVal dec As Long) As String
 
         Roman = ""
 
-        Dim intVal(12) As Integer
+        Dim intVal(12) As Long
         Dim strVal(12) As String, intLoopCounter As Integer
 
         intVal(0) = 1 : strVal(0) = "I"
@@ -723,7 +723,7 @@ Public Class SkillFunctions
         Return True
     End Function
 
-    Public Shared Function IsSkillTrained(ByVal cPilot As EveHQPilot, ByVal skillName As String, Optional ByVal toLevel As Integer = 0) As Boolean
+	Public Shared Function IsSkillTrained(ByVal cPilot As EveHQPilot, ByVal skillName As String, Optional ByVal toLevel As Long = 0) As Boolean
         If cPilot.PilotSkills.ContainsKey(skillName) Then
             Dim mySkill As EveHQPilotSkill = cPilot.PilotSkills(skillName)
             Dim myLevel As Integer = CInt(mySkill.Level)
